+++ apps/landing-onco/.eslintrc.js
@@ -1,0 +1,7 @@
+module.exports = {
+  root: true,
+  extends: ["custom"],
+  rules: {
+    "react-hooks/exhaustive-deps": 'off'
+  }
+};
