+++ apps/landing-onco/app/[lang]/favicon.ico
