+++ apps/landing-onco/components/cover/Cover.tsx
@@ -3,10 +3,9 @@
 import { useEffect, useRef } from "react";
 import { LandingButton } from "../shared/button/LandingButton";
 import { Molecules } from "./components/Molecules";
-import { mouseParallax } from "@/app/utils/animations/MouseParallax";
+import { mouseParallax } from "@utils/animations/MouseParallax";
 import { H1 } from "ui";
 import { openUrl } from "utils";
-import { LINKS } from "@/app/constants/links.const";
 
 export const Cover = () => {
   const containerRef = useRef<HTMLDivElement>(null);
