+++ apps/landing-onco/components/resources/Resources.tsx
