+++ apps/landing-onco/components/resources/components/Resource.tsx
