+++ apps/landing-onco/components/resources/components/ResourcePattern.tsx
