+++ apps/landing-onco/components/resources/types/resource.type.ts
