+++ apps/landing-onco/next.config.js
@@ -1,6 +1,6 @@
 /** @type {import('next').NextConfig} */
 module.exports = {
     reactStrictMode: true,
-    transpilePackages: ["ui"],
+    transpilePackages: ["ui", "i18n", "rest"],
   };
