+++ apps/landing-onco/package.json
@@ -18,6 +18,8 @@
     "gsap": "^3.12.2",
     "next": "13.4.19",
     "react": "18.2.0",
+    "negotiator": "^0.6.3",
+    "react-cookies": "^0.1.1",
     "react-dom": "18.2.0",
     "react-hook-form": "^7.46.2",
     "tsconfig": "*",
@@ -25,6 +27,7 @@
     "ui": "*"
   },
   "devDependencies": {
+    "@types/js-cookie": "^3.0.3",
     "autoprefixer": "^10.4.15",
     "postcss": "^8.4.28",
     "tailwindcss": "^3.3.3"
