+++ apps/web/tsconfig.json
@@ -4,7 +4,7 @@
     "plugins": [{ "name": "next" }],
     "baseUrl": ".",
     "paths": {
-      "@app/*": ["app/[lang]/*"],
+      "@*": ["app/[lang]/*"],
       "@core/*": ["core/*"],
       "@components/*": ["components/*"],
       "@api/*": ["api/*"],
